fullpush

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -10,7 +10,7 @@
  */
 void hbd(unsigned long int n, int i)
 {
-	if ((1 << i) > n) 
+	if ((1 << i) > (int)n) 
 		return;
 	hbd(n, i + 1);
 	printf("%d", (n & (1 << i)) ? 1 : 0);
